Backend/requirements.txt
bcrypt==4.0.1
cryptography==39.0.2
dnspython==2.3.0
fastapi==0.92.0
fastapi-login==1.8.3
jwt==1.3.1
kafka==1.3.5
kafka-python==2.0.2
mongoengine==0.27.0
motor==3.1.1
passlib==1.7.4
path==16.6.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
pymongo==4.3.3
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
regex==2023.3.23
typing_extensions==4.5.0
uvicorn==0.20.0
